Support default stations
The "Eagles Wing" location does not use stations to subdivide their
offerings. Wok now creates a default station with the ID of 0 when
this is the case and populates it with the correct menu information.
The default station is also instructed to ignore requests to fetch
data because it will not work.